add get query

+++ src/Services/Testing/Testing.API/Infrastructure/Models/PaginationResult.cs
@@ -5,14 +5,14 @@ public record PaginationResult<T>
     public int CurrentPage { get; }
     public int TotalPages { get; }
     public int PageSize { get; }
-    public int TotalCount { get; }
+    public long TotalCount { get; }
 
     public bool HasPrevious => CurrentPage > 1;
     public bool HasNext => CurrentPage < TotalPages;
 
     public ICollection<T> Rows { get; }
 
-    public PaginationResult(List<T> items, int count, int pageNumber, int pageSize)
+    public PaginationResult(List<T> items, long count, int pageNumber, int pageSize)
     {
         Rows = items;
         TotalCount = count;
